Skip illegal characters when creating the XML text
Tests can output any valid byte, but XML specification restricts the set of
valid chars to:

Char ::= #x9 | #xA | #xD | [#x20-#xD7FF] | [#xE000-#xFFFD] | [#x10000-#x10FFFF]

As we receive the output in a binary, the range starting in 16#100 is out of
scope. However, it is possible to receive an output containing illegal bytes in
the range 0x00..0x1F.

We skip those as it is not possible to escape them in any way.